Split `common.ts` into `tree-description` and `loader-utils`
tree-description: tree-related stuff, such as walking the tree and collapsing nodes
loader-tree: parsing-related stuff
